Split browser host registration into per-OS adapters
